cleanup after rebase

+++ mls-rs/src/group/application_data.rs
@@ -60,4 +60,4 @@ impl fmt::Debug for ComponentData {
     }
 }
 
-pub type ComponentId = u32;
+pub type ComponentId = u16;
